+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.12](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.11...@pyroscope/panel-plugin@1.3.12) (2022-07-01)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.11](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.10...@pyroscope/panel-plugin@1.3.11) (2022-07-01)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.3.10](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.9...@pyroscope/panel-plugin@1.3.10) (2022-06-30)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin